workaround bug in SecretKeyFactory
SecretKeyFactory had a bug before android-20/4.4 where it would only use
the first 8-bits of data in a unicode character.  In 4.4, they fixed the
bug, but that is not backwards compatible.  We need to stick with the old
behavior for now, until we figure out a migration procedure.

fixes #3664 https://dev.guardianproject.info/issues/3664
refs #3710 https://dev.guardianproject.info/issues/3710